Make check for image-should-fail quiet
We perform a check for this file after an image build fails to see if it
was intentional. This outputs a message which looks like the failure is
a result of the check when the file is missing which is very confusing
to users. Lets be more quiet about it.

Change-Id: I9b1f6c7c57021d11b4f91cd3c33846fae692b928